+++ parity-crypto/src/publickey/secret_key.rs
@@ -138,7 +138,7 @@ impl Secret {
 	pub fn dec(&mut self) -> Result<(), Error> {
 		match self.is_zero() {
 			true => {
-				*self = Secret::try_from(super::MINUS_ONE_KEY)
+				*self = Self::copy_from_slice(&super::MINUS_ONE_KEY)
 					.expect("Constructing a secret key from a known-good constant works; qed.");
 				Ok(())
 			}
@@ -213,14 +213,6 @@ impl Secret {
 	}
 }
 
-#[deprecated(since = "0.6.2", note = "please use `copy_from_str` instead, input is not zeroized")]
-impl FromStr for Secret {
-	type Err = Error;
-	fn from_str(s: &str) -> Result<Self, Self::Err> {
-		Ok(H256::from_str(s).map_err(|e| Error::Custom(format!("{:?}", e)))?.into())
-	}
-}
-
 impl From<[u8; 32]> for Secret {
 	#[inline(always)]
 	fn from(mut k: [u8; 32]) -> Self {
@@ -236,27 +228,6 @@ impl From<H256> for Secret {
 		let result = s.0.into();
 		s.0.zeroize();
 		result
-	}
-}
-
-#[deprecated(since = "0.6.2", note = "please use `copy_from_str` instead, input is not zeroized")]
-impl TryFrom<&str> for Secret {
-	type Error = Error;
-
-	fn try_from(s: &str) -> Result<Self, Error> {
-		s.parse().map_err(|e| Error::Custom(format!("{:?}", e)))
-	}
-}
-
-#[deprecated(since = "0.6.2", note = "please use `copy_from_slice` instead, input is not zeroized")]
-impl TryFrom<&[u8]> for Secret {
-	type Error = Error;
-
-	fn try_from(b: &[u8]) -> Result<Self, Error> {
-		if b.len() != SECP256K1_SECRET_KEY_SIZE {
-			return Err(Error::InvalidSecretKey);
-		}
-		Ok(Self { inner: Box::new(H256::from_slice(b)) })
 	}
 }