+++ parity-crypto/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog].
 - Bump `rust-secp256k1` to v0.19, always allow zero signatures. [#438](https://github.com/paritytech/parity-common/pull/438)
 - Updated `rlp` to 0.5. [#463](https://github.com/paritytech/parity-common/pull/463)
 - Updated dependencies. [#483](https://github.com/paritytech/parity-common/pull/483)
+- Remove deprecated trait impls `FromStr`/`TryFrom` for `Secret` [#495](https://github.com/paritytech/parity-common/pull/495)
 
 ## [0.6.2] - 2020-06-19
 - Put `Secret` memory on heap. [#400](https://github.com/paritytech/parity-common/pull/400)